Extract shared parser code into src/utilities/ and src/rotations.js
The parser modules under src/parsers/ accumulated airline-agnostic helpers — number/cents conversion, regex match helpers, date/timezone math, IATA→country lookup, indemnity amount lookup, and the rotation-building pipeline (buildRots/addIndemnities/mergeRots). With a second airline arriving, those helpers need to be reusable from any parser, not just AF's.

Split them by responsibility:

  src/utilities/    leaf utilities, no domain knowledge
    numbers.js     decimal/cents helpers and sum
    regex.js       matchAll / matchFirst / matchLast
    dates.js       iso2TZ, iso2FR, iso2AST, fr2iso, numberOfDays, diffHours, lastDayInMonthISO
    iata.js        iata2country (owns the airports.json import)
    indemnity.js   findAmount, findAmountEuros, AmountError

  src/rotations.js  tax/rotation domain pipeline
    buildRots, addIndemnities, mergeRots, mergeFlights, formula constants,
    monthsRotsIterator / arrayRotsIterator

The AF parsers (ep5Parser, payParser, nightsAFParser) now import from these modules instead of duplicating the logic. Components, stores, and tests are repointed at the new paths. The rotation tests move from test/ep5/ to test/rotations/ to match.

Pure refactor — no behavior change.
